+++ src/main/java/org/zalando/nakadi/service/EventTypeService.java
@@ -501,6 +501,11 @@ public class EventTypeService {
             final JSONObject schemaAsJson = new JSONObject(eventTypeSchema);
             final Schema schema = SchemaLoader.load(schemaAsJson);
 
+            if (schemaAsJson.has("type") && schemaAsJson.getString("type").equals("array")
+                    && eventType.getCategory() != EventCategory.UNDEFINED) {
+                throw new InvalidEventTypeException("\"type\" of root element in schema cannot be \"array\"");
+            }
+
             if (eventType.getCategory() == EventCategory.BUSINESS && schema.definesProperty("#/metadata")) {
                 throw new InvalidEventTypeException("\"metadata\" property is reserved");
             }